requirements.txt
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.10
django-appconf==1.0.3
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==2.0.4
-e git+https://github.com/michaelkuty/django-oscar-cash-on-delivery@d5a3c141bfdea143995bbcc89ba90c607c069be7#egg=django_oscar_cash_on_delivery
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.5
factory-boy==2.12.0
Faker==4.0.0
idna==2.9
phonenumbers==8.11.4
Pillow==7.0.0
purl==1.5
python-dateutil==2.8.1
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.9
